Revert "Revert "Sh product page handling""
