+++ src/App.css
@@ -1,15 +1,4 @@
-/* //דף בולקומרס שהצגת התמונות יהיה רספונסיבי */
-/* .heightRow{
-  height: 47vh;
-} */
-.Aheight a img{
-  height: 47vh;
-  border-image:1px red solid ;
-  background-size: contain;
-}
-.eye{
-  color: #707070;
-}
+
 /* //מבט מהיר */
 .sideBar{
   display: block;
@@ -37,132 +26,28 @@ font: normal normal normal 13px/15px Roboto;
 width: 100%;
 height: 35px;
 margin-bottom: 8vh;
-
-
 }
 
 .qu{
   display: inline-block;
 }
-/* //החלק של האינפוט נמבר של פלוס ומינוס  */
-.number-input input[type="number"] {
-  -webkit-appearance: textfield;
-    -moz-appearance: textfield;
-          appearance: textfield;
-}
-.number-input input[type=number]::-webkit-inner-spin-button,
-.number-input input[type=number]::-webkit-outer-spin-button {
-  -webkit-appearance: none;
-}
-.number-input {
-  margin-bottom: 1.5rem;
-  display: inline-block;
-  margin-left: 28%;
-}
-.number-input button {
-  -webkit-appearance: none;
-  background-color: transparent;
-  border: none;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  margin: 0;
-  position: relative;
-}
-.number-input button:before,
-.number-input button:after {
-  display: inline-block;
-  position: absolute;
-  content: '';
-  height: 2px;
-  transform: translate(-50%, -50%);
-}
-.number-input button.plus:after {
-  transform: translate(-50%, -50%) rotate(90deg);
-}
-.number-input input[type=number] {
-  text-align: center;
-}
-.number-input.number-input {
-  border: 1px solid #ced4da;
-  width:4.2rem;
-  border-radius: .25rem;
-  
-}
-.number-input.number-input button {
-  width: 1.3rem;
-  height: .7rem;
-}
-.number-input.number-input button.minus {
-  padding-left: 10px;
-}
-.number-input.number-input button:before,
-.number-input.number-input button:after {
-  width: .7rem;
-  background-color: #495057;
-}
-.number-input.number-input input[type=number] {
-  max-width: 1.4rem;
-  /* padding: .5rem; */
-  border: 1px solid #ced4da;
-  border-width: 0 1px;
-  font-size: 1rem;
-  height: 2rem;
-  color: #495057;
-}
-
-/* החלק של סיידבר שנפתח ונסגר מבט מהיר */
-.w3-sidebar{
-  height:90vh;
-  width:31%;
-  background-color:#fff;
-  /* position:fixed!important; */
-  position:absolute;
-  z-index:1;overflow:auto
-}
-.w3-bar-block .w3-bar-item{
-  width:100%;display:block;
-  padding:8px 16px;text-align:left;
-  border:none;white-space:normal;
-  float:none;outline:0}
-.w3-border-right{
-  border-right:1px solid #ccc!important}
-
-/* // פוטר דף הבית */
-.colorFooter{
-  background-color:#F29544 ;
-  color: #FFFFFF;
-  height: 40vh;
-}
-.size{
-font-size: 0.800em !important;
-}
-.BtnGoToShop{
-  border: 1px solid #707070;
-  color:black;
-  background-color:transparent;
-  height: 30px;
-  width:11%;
-  margin-left: 45%;
-  font-weight: bold;
-  font-size: 0.870em;
-}
+/* עד כאן החלק של מבט מהיר */
 .borderTop{
   border-top: 5px solid red;
   text-align: center;
   bottom: 0;
   position: fixed;
 }
-.logo{
-  height: 30px;
+.logoHeader{
+  height: 35px;
 }
 
-.footer {
+/* .footer {
   padding-bottom: 72px;
 }
 .footer__center {
   max-width: 1244px;
-}
+} */
 
 /* .center {
   padding: 0 32px;
@@ -229,8 +114,3 @@ select::-ms-expand {
 .field__select:focus {
   border-color: #5E81F4;
 } */
-/* //categoryBullcommerceNew */
-.titleCategory{
-  border-bottom: 1px solid black;
-  border-top: 1px solid black;
-}
