+++ src/assets/scss/now-ui-kit/_pagination.scss
@@ -63,8 +63,8 @@
             &,
             &:focus,
             &:hover{
-                background-color: $brand-info;
-                border-color: $brand-info;
+                background-color: orange;
+                border-color:ora;
             }
         }
     }
