+++ src/assets/scss/now-ui-kit/_variables.scss
@@ -147,7 +147,7 @@ $btn-icon-size-lg:              3.6rem         !default;
 $font-size-h1:                 3.5em        !default; // ~ 49px
 $font-size-h2:                 2.5em        !default; // ~ 35px
 $font-size-h3:                 1.825em      !default; // ~ 28px
-$font-size-h4:                 1.5em        !default; // ~ 24px
+$font-size-h4:                 1.2em        !default; // ~ 24px
 $font-size-h5:                 1.3em        !default; // ~ 22px
 $font-size-h6:                 .9em         !default; // ~ 12px
