+++ src/assets/scss/now-ui-kit/cards/_card-extend.scss
@@ -96,7 +96,7 @@
   }
 
   .card-title{
-      margin-top: 15px;
+      margin-top:-9%;
       line-height: 1.25em;
   }
