+++ src/assets/scss/now-ui-kit/cards/_card-product.scss
@@ -36,7 +36,7 @@
     .card-title,
     .category,
     .card-description{
-        text-align: center;
+        // text-align: center;
     }
     .category{
         margin-bottom: 0;
@@ -46,9 +46,9 @@
     }
 
     .price{
-        font-size: 16px;
-        color:black;
-        font-weight: bold;
+        font-size: 14px;
+        color:#666666;
+        // font-weight: bold;
 
     }
     .price-old{
