+++ src/components/navbars/ScrollTransparentNavbar.js
@@ -1,6 +1,7 @@
 import React from "react";
-import { Link } from "react-router-dom";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { Link } from "react-router-dom";
+import "./scrollNavbar.css"
 
 // reactstrap components
 import {
@@ -14,6 +15,8 @@ import {
   Nav,
   Container,
   UncontrolledTooltip,
+  Row,
+  Col,
 } from "reactstrap";
 
 import { connect } from 'react-redux';
@@ -23,9 +26,9 @@ import { connect } from 'react-redux';
 function ScrollTransparentNavbar(props) {
   // const [collapseOpen, setCollapseOpen] = React.useState(false);
   // const [navbarColor, setNavbarColor] = React.useState(
-    //   (document.documentElement.scrollTop > 499 || document.body.scrollTop) > 499
-    //     ? ""
-    //     : " navbar-transparent"
+  //   (document.documentElement.scrollTop > 499 || document.body.scrollTop) > 499
+  //     ? ""
+  //     : " navbar-transparent"
   // );
   // buyButtonColor
   const [, setBuyButtonColor] = React.useState(
@@ -60,13 +63,13 @@ function ScrollTransparentNavbar(props) {
     <>
 
       {/* מחקנו מבצע שההידר יהיה סטיקי */}
-      {/* className={"fixed-top" + navbarColor}  */}
+      {/* className={"fixed-top" + navbarColor} */}
       <Navbar color="white" expand="lg">
-        <Container>
+        <Container className="d-flex justify-content-between">
           <div className="navbar-translate">
             <NavbarBrand to="/" tag={Link} id="navbar-brand">
               <img alt="..."
-                src={props.objectFields.logo} className="logo"></img>
+                src={props.objectFields.logo} className="logoHeader"></img>
             </NavbarBrand>
             <UncontrolledTooltip target="navbar-brand">
               To Home Store Page
@@ -76,14 +79,13 @@ function ScrollTransparentNavbar(props) {
             // isOpen={collapseOpen}
             navbar>
             <Nav className="mx-auto" id="ceva" navbar>
-              {props.categories.filter(item => 
-              !item.masterCategory && item.masterCategory === null
-              // {
-              //   if (!item.masterCategory && item.masterCategory === null)
-              //     return item
-              // }
-              ).map((item, index) => (
+              {props.categories.filter(item => {
+                if (!item.masterCategory && item.masterCategory === null)
+                  return item
+              }).map((item, index) => (
+
                 <UncontrolledDropdown nav key={index}>
+
                   <DropdownToggle
                     caret
                     color="default"
@@ -93,25 +95,33 @@ function ScrollTransparentNavbar(props) {
                     nav
                     onClick={(e) => e.preventDefault()}
                   >
-                    <i
-                      aria-hidden={true}
-                      className="now-ui-icons files_paper"
-                    ></i>
                     <p>{item.categoryName}</p>
                   </DropdownToggle>
+
                   <DropdownMenu aria-labelledby="navbarDropdownMenuLink" right>
                     {item.childrenCategory.map((child, index) => (
                       <DropdownItem key={index} to="/sections#contact-us" tag={Link}>
-                        <i className="now-ui-icons tech_mobile"></i>
+                        {/* <i className="now-ui-icons tech_mobile"></i> */}
                         {child.categoryName}
                       </DropdownItem>
                     ))}
                   </DropdownMenu>
                 </UncontrolledDropdown>
+
               ))}
             </Nav>
-            <FontAwesomeIcon className="mt-2 ml-5" icon={['fas', 'search']}></FontAwesomeIcon>
-            <FontAwesomeIcon className="mt-2 ml-3" icon={['fas', 'shopping-cart']}></FontAwesomeIcon>
+            <Link to="/">
+              <FontAwesomeIcon className="mt-2 mr-3" icon={['fas', 'search']}></FontAwesomeIcon>
+            </Link>
+            <div className="separatorStripe"></div>
+            <Link to={"/" + props.objectFields.storeName + "/cart"}>
+              <FontAwesomeIcon className="mt-2 ml-3" icon={['fas', 'shopping-cart']}></FontAwesomeIcon>
+            </Link>
+
+            <Link to={"/" + props.objectFields.storeName + "/"}>
+
+              <FontAwesomeIcon className="mt-2 ml-3" icon={['far', 'user-circle']}></FontAwesomeIcon>
+            </Link>
           </Collapse>
         </Container>
       </Navbar>
