+++ src/components/navbars/scrollNavbar.css
@@ -1,0 +1,5 @@
+.separatorStripe{
+    width: 2px;
+    height: 20px;
+    background-color: black;
+}
