+++ src/components/store design/bullcommerce.css
@@ -1,0 +1,62 @@
+.heightViewProduct{
+    height: 47vh;
+    border-image:1px red solid ;
+  
+  }
+  .Aheight a img{
+    height: 47vh;
+    border-image:1px red solid ;
+    background-size: contain;
+  }
+  .eye{
+    color: #707070;
+  }
+/* הידר תמונת פתיחה */
+
+  .arrowIcon{
+    color:black;
+    height: 40px;
+    width:40px
+  
+  }
+
+  /* // פוטר דף הבית */
+.colorFooter{
+    background-color:#F29544 ;
+    color: #FFFFFF;
+    height: 40vh;
+  }
+  .size{
+  font-size: 0.800em !important;
+  }
+  
+ /* //////// */
+ .BtnGoToShop{
+    border: 1px solid #707070;
+    color:black;
+    background-color:transparent;
+    height: 30px;
+    width:11%;
+    margin-left: 45%;
+    font-weight: bold;
+    font-size: 0.870em;
+  } 
+  /* //// */
+  /* החלק של סיידבר שנפתח ונסגר מבט מהיר */
+.w3-sidebar{
+    height:90vh;
+    width:31%;
+    background-color:#fff;
+    /* position:fixed!important; */
+    position:absolute;
+    z-index:1;overflow:auto
+  }
+  .w3-bar-block .w3-bar-item{
+    width:100%;display:block;
+    padding:8px 16px;text-align:left;
+    border:none;white-space:normal;
+    float:none;outline:0}
+  .w3-border-right{
+    border-right:1px solid #ccc!important}
+  
+
