+++ src/components/store design/categoryBullcommerce.css
@@ -1,0 +1,47 @@
+.titleCategory{
+    border-bottom: 1px solid #dbd7d7;
+    border-bottom-width: 2px;
+    border-top: 1px solid#dbd7d7;
+    border-top-width: 2px;
+    color:#999999;
+    font-weight: bold;
+   
+  }
+  .iconGridAndList{
+     color:black 
+  }
+  .pagerCategory{
+    border-top: 1px solid#dbd7d7;
+    font-weight: bold;
+
+  }
+.frameToProductView{
+    border: 1px solid aqua;
+    width: 170px;
+    height: 245px;
+    /* במסך הגדול זה היה כך */
+    /* width:200px;
+    height: 175px; */
+
+}
+.imageProduct{
+    object-fit: cover;
+    object-position: 50% 50%;
+width:100%;
+height:100%;
+}
+
+.card-title{
+    font-size: 14px;
+    margin-bottom: -0.12rem !important;
+}
+.card-description{
+    font-size: 12px;
+    margin-bottom: -0.5rem !important;
+}
+.price{
+    font-size: 12px !important;
+}
+.as{
+    text-decoration: underline overline
+}
