+++ src/components/store design/filteredProducts.js
@@ -29,6 +29,37 @@ import { useHistory } from "react-router-dom";
 
 function FilteredProducts(props) {
 
+        // 
+
+        // React.useEffect(() => {
+        //         if (
+        //                 !document
+        //                         .getElementById("sliderRegular")
+        //                         .classList.contains("noUi-target")
+        //         ) {
+        //                 Slider.create(document.getElementById("sliderRegular"), {
+        //                         start: [50],
+        //                         connect: [true, false],
+        //                         step: 0.5,
+        //                         range: { min: 0, max: 100 },
+        //                 });
+        //         }
+        //         if (
+        //                 !document.getElementById("sliderDouble").classList.contains("noUi-target")
+        //         ) {
+        //                 Slider.create(document.getElementById("sliderDouble"), {
+        //                         start: [20, 80],
+        //                         connect: [false, true, false],
+        //                         step: 1,
+        //                         range: { min: 0, max: 100 },
+        //                 });
+        //         }
+        // });
+
+        // 
+
+
+
         // collapse states and functions
         const [collapses, setCollapses] = React.useState([1]);
         const changeCollapse = (collapse) => {
@@ -40,33 +71,33 @@ function FilteredProducts(props) {
         };
 
         const history = useHistory();
+        // React.useEffect(() => {
 
-        React.useEffect(() => {
 
-                let min = props.storeProducts[0].price,
-                        max = props.storeProducts[0].price;
-                props.products.forEach(product => {
-                        if (product.price > max)
-                                max = product.price;
-                        if (product.price < min)
-                                min = product.price;
-                });
+        //         let min = props.storeProducts[0].price,
+        //                 max = props.storeProducts[0].price;
+        //         props.storeProducts.forEach(product => {
+        //                 if (product.price > max)
+        //                         max = product.price;
+        //                 if (product.price < min)
+        //                         min = product.price;
+        //         });
 
-                if (
-                        !document.getElementById("sliderRefine").classList.contains("noUi-target")
-                ) {
-                        Slider.create(document.getElementById("sliderRefine"), {
-                                start: [min + 20, max],
-                                connect: [false, true, false],
-                                step: 1,
-                                range: { min: min, max: max },
-                        }).on("update", function (values) {
+        //         if (
+        //                 !document.getElementById("sliderRefine").classList.contains("noUi-target")
+        //         ) {
+        //                 Slider.create(document.getElementById("sliderRefine"), {
+        //                         start: [min + 20, max],
+        //                         connect: [false, true, false],
+        //                         step: 1,
+        //                         range: { min: min, max: max },
+        //                 }).on("update", function (values) {
 
-                                props.setSliderMin(Math.round(values[0]));
-                                props.setSliderMax(Math.round(values[1]));
-                        });
-                }
-        }, []);
+        //                         props.setSliderMin(Math.round(values[0]));
+        //                         props.setSliderMax(Math.round(values[1]));
+        //                 });
+        //         }
+        // }, []);
 
         const onFilter = () => {
 
@@ -79,18 +110,115 @@ function FilteredProducts(props) {
 
         return (
                 <>
-                        {/* <ScrollTransparentNavbar /> */}
-                        {/* <div className="wrapper"> */}
-                        {/* <EcommerceHeader /> */}
-                        {/* <div className="main">
-                                        <div className="section">
-                                                <Container> */}
-                        {/* <h2 className="section-title">Find what you need</h2> */}
-                        {/* <Row> */}
-                        {/* <Col md="3">
-                                                                        <div className="collapse-panel"> */}
                         <CardBody>
                                 <Card className="card-refine card-plain">
+                                        <CardHeader id="headingTwo" role="tab">
+                                                <h6>
+                                                        <a
+                                                                className="text-info as"
+                                                                aria-expanded={collapses.includes(2)}
+                                                                data-toggle="collapse"
+                                                                data-parent="#accordion"
+                                                                href="#pablo"
+                                                                onClick={(e) => {
+                                                                        e.preventDefault();
+                                                                        changeCollapse(2);
+                                                                }}
+                                                        >
+                                                                Categories{" "}
+                                                                <FontAwesomeIcon icon={['fas', 'chevron-down']}></FontAwesomeIcon>
+
+                                                                {/* <i className="now-ui-icons arrows-1_minimal-down"></i> */}
+                                                        </a>
+                                                </h6>
+                                        </CardHeader>
+                                        <Collapse isOpen={collapses.includes(2)}>
+                                                <CardBody>
+                                                        {props.categories.map((item, index) => (
+                                                                <FormGroup check key={index}>
+                                                                        <Label check>
+                                                                                <Input type="checkbox"></Input>
+                                                                                <span className="form-check-sign"></span>
+                                                                                {item.categoryName}
+                                                                        </Label>
+                                                                </FormGroup>
+                                                        ))}
+                                                </CardBody>
+                                        </Collapse>
+
+                                </Card>
+                                <Card className="card-refine card-plain">
+                                        <CardHeader id="headingfour" role="tab">
+                                                <h6>
+                                                        <a
+                                                                className="text-info"
+                                                                aria-expanded={collapses.includes(4)}
+                                                                data-toggle="collapse"
+                                                                data-parent="#accordion"
+                                                                href="#pablo"
+                                                                onClick={(e) => {
+                                                                        e.preventDefault();
+                                                                        changeCollapse(4);
+                                                                }}
+                                                        >
+                                                                Color{" "}
+                                                                <FontAwesomeIcon icon={['fas', 'chevron-down']}></FontAwesomeIcon>
+
+                                                                {/* <i className="now-ui-icons arrows-1_minimal-down"></i> */}
+                                                        </a>
+                                                </h6>
+                                        </CardHeader>
+                                        <Collapse isOpen={collapses.includes(4)}>
+                                                <CardBody>
+                                                        <FormGroup check>
+                                                                <Label check>
+                                                                        <Input type="checkbox"></Input>
+                                                                        <span className="form-check-sign"></span>
+                                Black
+                              </Label>
+                                                        </FormGroup>
+                                                        <FormGroup check>
+                                                                <Label check>
+                                                                        <Input type="checkbox"></Input>
+                                                                        <span className="form-check-sign"></span>
+                                Blue
+                              </Label>
+                                                        </FormGroup>
+                                                        <FormGroup check>
+                                                                <Label check>
+                                                                        <Input type="checkbox"></Input>
+                                                                        <span className="form-check-sign"></span>
+                                Brown
+                              </Label>
+                                                        </FormGroup>
+                                                        <FormGroup check>
+                                                                <Label check>
+                                                                        <Input type="checkbox"></Input>
+                                                                        <span className="form-check-sign"></span>
+                                Gray
+                              </Label>
+                                                        </FormGroup>
+                                                        <FormGroup check>
+                                                                <Label check>
+                                                                        <Input type="checkbox"></Input>
+                                                                        <span className="form-check-sign"></span>
+                                Green
+                              </Label>
+                                                        </FormGroup>
+                                                        <FormGroup check>
+                                                                <Label check>
+                                                                        <Input type="checkbox"></Input>
+                                                                        <span className="form-check-sign"></span>
+                                Purple
+                              </Label>
+                                                        </FormGroup>
+                                                </CardBody>
+                                        </Collapse>
+                                </Card>
+
+
+                                <Card className="card-refine card-plain">
+
                                         {/* <CardTitle tag="h4">
                                                 Refine{" "}
                                                 <Button
@@ -135,8 +263,8 @@ function FilteredProducts(props) {
                                                                 €{props.slideMin}
                                                         </span>
                                                         <span
-                                                                className="price-right pull-right"
-                                                                id="price-right"
+                                                        // className="price-right pull-right"
+                                                        // id="price-right"
                                                         >
                                                                 €{props.slideMax}
                                                         </span>
@@ -148,41 +276,7 @@ function FilteredProducts(props) {
                                                 </CardBody>
                                         </Collapse>
                                 </Card>
-                                <Card className="card-refine card-plain">
-                                        <CardHeader id="headingTwo" role="tab">
-                                                <h6>
-                                                        <a
-                                                                className="text-info"
-                                                                aria-expanded={collapses.includes(2)}
-                                                                data-toggle="collapse"
-                                                                data-parent="#accordion"
-                                                                href="#pablo"
-                                                                onClick={(e) => {
-                                                                        e.preventDefault();
-                                                                        changeCollapse(2);
-                                                                }}
-                                                        >
-                                                                Categories{" "}
-                                                                <FontAwesomeIcon icon={['fas', 'chevron-down']}></FontAwesomeIcon>
 
-                                                                {/* <i className="now-ui-icons arrows-1_minimal-down"></i> */}
-                                                        </a>
-                                                </h6>
-                                        </CardHeader>
-                                        <Collapse isOpen={collapses.includes(2)}>
-                                                <CardBody>
-                                                        {props.categories.map((item, index) => (
-                                                                <FormGroup check key={index}>
-                                                                        <Label check>
-                                                                                <Input type="checkbox"></Input>
-                                                                                <span className="form-check-sign"></span>
-                                                                                {item.categoryName}
-                                                                        </Label>
-                                                                </FormGroup>
-                                                        ))}
-                                                </CardBody>
-                                        </Collapse>
-                                </Card>
                                 <Card className="card-refine card-plain">
                                         <CardHeader id="headingThree" role="tab">
                                                 <h6>
@@ -272,74 +366,18 @@ function FilteredProducts(props) {
                                                 </CardBody>
                                         </Collapse>
                                 </Card>
-                                <Card className="card-refine card-plain">
-                                        <CardHeader id="headingfour" role="tab">
-                                                <h6>
-                                                        <a
-                                                                className="text-info"
-                                                                aria-expanded={collapses.includes(4)}
-                                                                data-toggle="collapse"
-                                                                data-parent="#accordion"
-                                                                href="#pablo"
-                                                                onClick={(e) => {
-                                                                        e.preventDefault();
-                                                                        changeCollapse(4);
-                                                                }}
-                                                        >
-                                                                Color{" "}
-                                                                <FontAwesomeIcon icon={['fas', 'chevron-down']}></FontAwesomeIcon>
+                                {/*  */}
+                                {/* <Col lg="3" sm="6">
+                                        <p className="category">Sliders</p>
+                                        <div className="slider" id="sliderRegular"></div>
+                                        <br></br>
+                                        <div className="slider slider-info" id="sliderDouble"></div>
+                                </Col> */}
 
-                                                                {/* <i className="now-ui-icons arrows-1_minimal-down"></i> */}
-                                                        </a>
-                                                </h6>
-                                        </CardHeader>
-                                        <Collapse isOpen={collapses.includes(4)}>
-                                                <CardBody>
-                                                        <FormGroup check>
-                                                                <Label check>
-                                                                        <Input type="checkbox"></Input>
-                                                                        <span className="form-check-sign"></span>
-                                Black
-                              </Label>
-                                                        </FormGroup>
-                                                        <FormGroup check>
-                                                                <Label check>
-                                                                        <Input type="checkbox"></Input>
-                                                                        <span className="form-check-sign"></span>
-                                Blue
-                              </Label>
-                                                        </FormGroup>
-                                                        <FormGroup check>
-                                                                <Label check>
-                                                                        <Input type="checkbox"></Input>
-                                                                        <span className="form-check-sign"></span>
-                                Brown
-                              </Label>
-                                                        </FormGroup>
-                                                        <FormGroup check>
-                                                                <Label check>
-                                                                        <Input type="checkbox"></Input>
-                                                                        <span className="form-check-sign"></span>
-                                Gray
-                              </Label>
-                                                        </FormGroup>
-                                                        <FormGroup check>
-                                                                <Label check>
-                                                                        <Input type="checkbox"></Input>
-                                                                        <span className="form-check-sign"></span>
-                                Green
-                              </Label>
-                                                        </FormGroup>
-                                                        <FormGroup check>
-                                                                <Label check>
-                                                                        <Input type="checkbox"></Input>
-                                                                        <span className="form-check-sign"></span>
-                                Purple
-                              </Label>
-                                                        </FormGroup>
-                                                </CardBody>
-                                        </Collapse>
-                                </Card>
+                                {/*  */}
+
+
+
                                 <Button
                                         className="btn-round"
                                         color="info"
@@ -357,6 +395,7 @@ function FilteredProducts(props) {
         );
 }
 
+
 export default connect(
         (state) => {
                 return {
@@ -373,8 +412,6 @@ export default connect(
                         setSliderMax: (x) => { dispatch(actions.setMaxPrice(x)) },
                         setFilteredItems: (x) => { dispatch(actions.setFilteredItems(x)) }
                 }
-        }
-)(FilteredProducts);
-
+        })(FilteredProducts);
