+++ src/components/store design/product_page/product.css
@@ -1,0 +1,79 @@
+#titleAndContent{
+    font-size:18px;
+    padding-left: 7% !important;
+}
+
+#titleAndContent h4{
+    font-size:18px;
+    margin-top: 5px;
+    font-style: bold;
+}
+
+.productColors .color{
+    float:left;
+    width: 37px;
+    height: 37px;
+    border-radius: 50%;
+    margin-right: 5px;
+}
+.productColors .color:last-child{
+    float: none;
+}
+
+#productImage{
+    border: 2px solid #707070;
+    align-items:center;
+}
+.productPrice span{
+    font-weight:900 ;
+}
+.productPrice del{
+    opacity: 0.3;
+    margin-right: 40px;
+}
+
+.addToCart{
+    border-color: #707070 !important;
+    color: #F29544 !important;
+    margin-left:10px !important;
+    width: 10em;
+}
+
+.addToCart:hover{
+    border-color: #F29544 !important;
+    background-color: #F29544 !important;
+    color: white !important;
+    
+}
+
+#titleAndContent .d{
+    border-top: 2px solid #959595;
+    /* height:20vh; */
+    padding-bottom: 1rem;
+    padding-top:1rem;
+}
+/* ggggggggggg */
+#moreDetails h3{
+    font-size: 20px;
+    color:#707070;
+    width:100%;
+    border-bottom: 2px solid #9B9B9B;
+}
+#titleSimilar{
+    text-align: center;
+    color:#707070;
+}
+.arrows{
+    text-align: left;
+}
+.similarImg{
+    background-color: #F5F5F5;
+    /* margin-right: 8px;
+    margin-top: 10px;
+    align-items:center; */
+}
+.similarImg img{
+    align-items: center;
+}
+
+
